FitnessCenter/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
exceptiongroup==1.2.2
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
kafka-python==2.0.2
packaging==24.1
pluggy==1.5.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
